Remove mock outputs and add quiet parameter
This commit implements the `quiet` parameter of runner that was
requested to be ported to the SDK. I looked through the code
and found only output to stdout that was used for development.
Removed it.

Quiet option in jobs can be used in the future to decide if
informal output should be done at all.